Get markdown-formatted full text of a partition.

The method is the worker behind the read-method, which will be called usually to reconstruct the full text of a partition and read it. The as.markdown-method can be customized for different classes inheriting from the partition-class.

as.markdown(.Object, ...)

# S4 method for partition
as.markdown(.Object, meta = getOption("polmineR.meta"),
  template = get_template(.Object), cpos = TRUE, cutoff = NULL,
  verbose = FALSE, ...)

# S4 method for plpr_partition
as.markdown(.Object, meta = NULL,
  template = get_template(.Object), cpos = FALSE, interjections = TRUE,
  cutoff = NULL, ...)

Arguments

.Object
The object to be converted, a partition, or a class inheriting from partition, such as plpr_partition.

...
further arguments

meta
The metainformation (s-attributes) to be displayed.

template
A template for formating output.

cpos
A logical value, whether to add cpos as ids in span elements.

cutoff
The maximum number of tokens to reconstruct, to avoid that full text is excessively long.

verbose
A logical value, whether to output messages.

interjections
A logical value, whether to format interjections.

Examples

use("polmineR")

#> ... activating corpus: germaparlmini

#> ... activating corpus: reuters

P <- partition("REUTERS", places = "argentina")

#> ... get encoding: latin1

#> ... get cpos and strucs

as.markdown(P)

#> [1] "## Corpus: REUTERS\n\n###  \n\n <span id=\"3997\" token=\"Argentine\">Argentine</span> <span id=\"3998\" token=\"crude\">crude</span> <span id=\"3999\" token=\"oil\">oil</span> <span id=\"4000\" token=\"production\">production</span> <span id=\"4001\" token=\"was\">was</span> <span id=\"4002\" token=\"down\">down</span> <span id=\"4003\" token=\"10.8\">10.8</span> <span id=\"4004\" token=\"pct\">pct</span> <span id=\"4005\" token=\"in\">in</span> <span id=\"4006\" token=\"January\">January</span> <span id=\"4007\" token=\"1987\">1987</span> <span id=\"4008\" token=\"to\">to</span> <span id=\"4009\" token=\"12.32\">12.32</span> <span id=\"4010\" token=\"mln\">mln</span> <span id=\"4011\" token=\"barrels\">barrels</span> <span id=\"4012\" token=\"from\">from</span> <span id=\"4013\" token=\"13.81\">13.81</span> <span id=\"4014\" token=\"mln\">mln</span> <span id=\"4015\" token=\"barrels\">barrels</span> <span id=\"4016\" token=\"in\">in</span> <span id=\"4017\" token=\"January\">January</span> <span id=\"4018\" token=\"1986\">1986</span> <span id=\"4019\" token=\"Yacimientos\">Yacimientos</span> <span id=\"4020\" token=\"Petroliferos\">Petroliferos</span> <span id=\"4021\" token=\"Fiscales\">Fiscales</span> <span id=\"4022\" token=\"said\">said</span> <span id=\"4023\" token=\"January\">January</span> <span id=\"4024\" token=\"1987\">1987</span> <span id=\"4025\" token=\"natural\">natural</span> <span id=\"4026\" token=\"gas\">gas</span> <span id=\"4027\" token=\"output\">output</span> <span id=\"4028\" token=\"totalled\">totalled</span> <span id=\"4029\" token=\"1.15\">1.15</span> <span id=\"4030\" token=\"billion\">billion</span> <span id=\"4031\" token=\"cubic\">cubic</span> <span id=\"4032\" token=\"metrers\">metrers</span> <span id=\"4033\" token=\"3.6\">3.6</span> <span id=\"4034\" token=\"pct\">pct</span> <span id=\"4035\" token=\"higher\">higher</span> <span id=\"4036\" token=\"than\">than</span> <span id=\"4037\" token=\"1.11\">1.11</span> <span id=\"4038\" token=\"billion\">billion</span> <span id=\"4039\" token=\"cubic\">cubic</span> <span id=\"4040\" token=\"metres\">metres</span> <span id=\"4041\" token=\"produced\">produced</span> <span id=\"4042\" token=\"in\">in</span> <span id=\"4043\" token=\"January\">January</span> <span id=\"4044\" token=\"1986\">1986</span> <span id=\"4045\" token=\"Yacimientos\">Yacimientos</span> <span id=\"4046\" token=\"Petroliferos\">Petroliferos</span> <span id=\"4047\" token=\"Fiscales\">Fiscales</span> <span id=\"4048\" token=\"added\">added</span> <span id=\"4049\" token=\"Reuter\">Reuter</span> \n"

as.markdown(P, meta = c("id", "places"))

#> [1] "## Corpus: REUTERS\n\n###  id: 708 // places: argentina\n\n <span id=\"3997\" token=\"Argentine\">Argentine</span> <span id=\"3998\" token=\"crude\">crude</span> <span id=\"3999\" token=\"oil\">oil</span> <span id=\"4000\" token=\"production\">production</span> <span id=\"4001\" token=\"was\">was</span> <span id=\"4002\" token=\"down\">down</span> <span id=\"4003\" token=\"10.8\">10.8</span> <span id=\"4004\" token=\"pct\">pct</span> <span id=\"4005\" token=\"in\">in</span> <span id=\"4006\" token=\"January\">January</span> <span id=\"4007\" token=\"1987\">1987</span> <span id=\"4008\" token=\"to\">to</span> <span id=\"4009\" token=\"12.32\">12.32</span> <span id=\"4010\" token=\"mln\">mln</span> <span id=\"4011\" token=\"barrels\">barrels</span> <span id=\"4012\" token=\"from\">from</span> <span id=\"4013\" token=\"13.81\">13.81</span> <span id=\"4014\" token=\"mln\">mln</span> <span id=\"4015\" token=\"barrels\">barrels</span> <span id=\"4016\" token=\"in\">in</span> <span id=\"4017\" token=\"January\">January</span> <span id=\"4018\" token=\"1986\">1986</span> <span id=\"4019\" token=\"Yacimientos\">Yacimientos</span> <span id=\"4020\" token=\"Petroliferos\">Petroliferos</span> <span id=\"4021\" token=\"Fiscales\">Fiscales</span> <span id=\"4022\" token=\"said\">said</span> <span id=\"4023\" token=\"January\">January</span> <span id=\"4024\" token=\"1987\">1987</span> <span id=\"4025\" token=\"natural\">natural</span> <span id=\"4026\" token=\"gas\">gas</span> <span id=\"4027\" token=\"output\">output</span> <span id=\"4028\" token=\"totalled\">totalled</span> <span id=\"4029\" token=\"1.15\">1.15</span> <span id=\"4030\" token=\"billion\">billion</span> <span id=\"4031\" token=\"cubic\">cubic</span> <span id=\"4032\" token=\"metrers\">metrers</span> <span id=\"4033\" token=\"3.6\">3.6</span> <span id=\"4034\" token=\"pct\">pct</span> <span id=\"4035\" token=\"higher\">higher</span> <span id=\"4036\" token=\"than\">than</span> <span id=\"4037\" token=\"1.11\">1.11</span> <span id=\"4038\" token=\"billion\">billion</span> <span id=\"4039\" token=\"cubic\">cubic</span> <span id=\"4040\" token=\"metres\">metres</span> <span id=\"4041\" token=\"produced\">produced</span> <span id=\"4042\" token=\"in\">in</span> <span id=\"4043\" token=\"January\">January</span> <span id=\"4044\" token=\"1986\">1986</span> <span id=\"4045\" token=\"Yacimientos\">Yacimientos</span> <span id=\"4046\" token=\"Petroliferos\">Petroliferos</span> <span id=\"4047\" token=\"Fiscales\">Fiscales</span> <span id=\"4048\" token=\"added\">added</span> <span id=\"4049\" token=\"Reuter\">Reuter</span> \n"

if (interactive()) read(P, meta = c("id", "places"))