Analyze context of a node word.

Retrieve the word context of a token, optionally checking for boundaries of a XML region.

# S4 method for partition
context(.Object, query, cqp = is.cqp,
  left = getOption("polmineR.left"), right = getOption("polmineR.right"),
  pAttribute = getOption("polmineR.pAttribute"), sAttribute = NULL,
  stoplist = NULL, positivelist = NULL, regex = FALSE, count = TRUE,
  mc = getOption("polmineR.mc"), verbose = TRUE, progress = TRUE)

# S4 method for character
context(.Object, query,
  pAttribute = getOption("polmineR.pAttribute"), sAttribute = NULL, ...)

# S4 method for partitionBundle
context(.Object, query, verbose = TRUE, ...)

# S4 method for cooccurrences
context(.Object, query, complete = FALSE)

# S4 method for Corpus
cooccurrences(.Object, query,
  pAttribute = getOption("polmineR.pAttribute"), ...)

Arguments

.Object
a partition or a partitionBundle object

query
query, which may by a character vector or a cqpQuery object

cqp
defaults to is.cqp-function, or provide TRUE/FALSE

left
no of tokens and to the left of the node word

right
no of tokens to the right of the node word

pAttribute
p-attribute of the query

sAttribute
if provided, it will be checked that corpus positions do not extend beyond the region defined by the s-attribute

stoplist
exclude a query hit from analysis if stopword(s) is/are in context. See positivelist for further explanation.

positivelist
character vector or numeric/integer vector: include a query hit only if token in positivelist is present. If positivelist is a character vector, it may include regular expressions (see parameter regex)

regex
logical, defaults to FALSE - whether stoplist and/or positivelist are regular expressions

count
logical

mc
whether to use multicore; if NULL (default), the function will get the value from the options

verbose
report progress, defaults to TRUE

progress
logical, whether to show progress bar

...
further parameters

complete
enhance completely

Value

depending on whether a partition or a partitionBundle serves as input, the return will be a context object, or a contextBundle object

Details

For formulating the query, CPQ syntax may be used (see examples). Statistical tests available are log-likelihood, t-test, pmi.

Examples

# NOT RUN {
  use("polmineR.sampleCorpus")
  p <- partition("PLPRBTTXT", list(text_type="speech"))
  y <- context(p, query = "Integration", pAttribute = "word")
  y <- context(p, query = "Integration", pAttribute = "word", positivelist = "Bildung")
  y <- context(
    p, query = "Integration", pAttribute = "word",
    positivelist = c("[aA]rbeit.*", "Ausbildung"), regex = TRUE
    )
# }