Get cooccurrence statistics.

Get cooccurrence statistics.

cooccurrences(.Object, ...)

# S4 method for character
cooccurrences(.Object, query, cqp = is.cqp,
  p_attribute = getOption("polmineR.p_attribute"), s_attribute = NULL,
  left = getOption("polmineR.left"), right = getOption("polmineR.right"),
  stoplist = NULL, positivelist = NULL, regex = FALSE, keep = NULL,
  cpos = NULL, method = "ll", mc = getOption("polmineR.mc"),
  verbose = FALSE, progress = FALSE, ...)

# S4 method for partition
cooccurrences(.Object, query, cqp = is.cqp,
  left = getOption("polmineR.left"), right = getOption("polmineR.right"),
  p_attribute = getOption("polmineR.p_attribute"), s_attribute = NULL,
  stoplist = NULL, positivelist = NULL, keep = NULL, method = "ll",
  mc = FALSE, progress = TRUE, verbose = FALSE, ...)

# S4 method for context
cooccurrences(.Object, method = "ll", verbose = FALSE)

# S4 method for Corpus
cooccurrences(.Object, query,
  p_attribute = getOption("polmineR.p_attribute"), ...)

# S4 method for partition_bundle
cooccurrences(.Object, query,
  mc = getOption("polmineR.mc"), ...)

Arguments

.Object
a partition object, or a character vector with a CWB corpus

...
further parameters that will be passed into bigmatrix (applies only of big=TRUE)

query
query, may by a character vector to match a token, or a CQP query

cqp
defaults to is.cqp-function, or provide TRUE/FALSE, relevant only if query is not NULL

p_attribute
the p-attribute of the tokens/the query

s_attribute
if provided, it will be checked that cpos do not extend beyond the region defined by the s-attribute

left
no of tokens and to the left of the node word

right
no of tokens to the right of the node word

stoplist
exclude a query hit from analysis if stopword(s) is/are in context (relevant only if query is nut NULL)

positivelist
character vector or numeric vector: include a query hit only if token in positivelist is present. If positivelist is a character vector, it is assumed to provide regex expressions (incredibly long if the list is long) (relevant only if query is nut NULL)

regex
logical, whether stoplist/positivelist are dealt with as regular expressions

keep
list with tokens to keep

cpos
integer vector with corpus positions, defaults to NULL - then the corpus positions for the whole corpus will be used

method
statistical test to use (defaults to "ll")

mc
whether to use multicore

verbose
logical, whether to be verbose

progress
logical, whether to be verbose

Value

a cooccurrences-class object

References

Baker, Paul (2006): Using Corpora in Discourse Analysis. London: continuum, p. 95-120 (ch. 5).

Manning, Christopher D.; Schuetze, Hinrich (1999): Foundations of Statistical Natural Language Processing. MIT Press: Cambridge, Mass., pp. 151-189 (ch. 5).

Examples

use("polmineR")

#> ... activating corpus: germaparlmini

#> ... activating corpus: reuters

merkel <- partition("GERMAPARLMINI", interjection = "speech", speaker = ".*Merkel", regex = TRUE)

#> ... get encoding: latin1

#> ... get cpos and strucs

merkel <- enrich(merkel, p_attribute = "word")

#> ... getting counts for p-attribute(s): word

cooc <- cooccurrences(merkel, query = "Deutschland")