Cooccurrences class.

S4 class to organize information of context analysis

# S4 method for cooccurrences
show(object)

# S4 method for cooccurrences_bundle
as.data.frame(x)

# S4 method for cooccurrences
format(x, digits = 2L)

# S4 method for cooccurrences
view(.Object)

# S4 method for cooccurrences_reshaped
view(.Object)

Arguments

object
object to work with

x
object to work with

digits
Integer indicating the number of decimal places (round) or significant digits (signif) to be used.

.Object
object to work with

Slots

call

Object of class character the call that generated the object

partition

Object of class character the partition the analysis is based on

size_partition

Object of class integer the size of the partition

left

Object of class integer number of tokens to the left.

right

Object of class integer number of tokens to the right.

p_attribute

Object of class character p-attribute of the query

corpus

Object of class character the CWB corpus used

stat

Object of class data.table statistics of the analysis

encoding

Object of class character encoding of the corpus

method

Object of class character statistical test(s) used