Get corpus/corpora available or used.

Calling corpus() will return a data.frame listing the corpora described in the active registry directory, and some basic information on the corpora. If object is an object inheriting from the textstat, or the bundle class, the corpus used to generate the object is returned.

corpus(object)

# S4 method for textstat
corpus(object)

# S4 method for bundle
corpus(object)

# S4 method for missing
corpus()

Arguments

object
An object inheriting from the textstat or bundle superclasses.

Examples

use("polmineR")

#> ... activating corpus: germaparlmini

#> ... activating corpus: reuters

corpus()

#>           corpus      size template
#> 1         ARENEN  28842158     TRUE
#> 2     AUSTROPARL  70308143     TRUE
#> 3          BTPDF 139526576     TRUE
#> 4  GERMAPARLMINI    222201     TRUE
#> 5       KEYWORDS  44734138     TRUE
#> 6    KEYWORDS_DE   7525003     TRUE
#> 7    KEYWORDS_EN  36956586     TRUE
#> 8    KEYWORDS_FR  18904184     TRUE
#> 9   KEYWORDS_FR2  18904184     TRUE
#> 10         NEXIS 164740428     TRUE
#> 11        PLPRBT       -11     TRUE
#> 12     PLPRBTTXT  83405660     TRUE
#> 13       PLPRTXT  83195275     TRUE
#> 14    REPUBBLICA  26270819     TRUE
#> 15       REUTERS      4050     TRUE
#> 16          UNGA  43072609     TRUE
#> 17      WPRGAGGR   8913223     TRUE
#> 18          ZEIT 327839748     TRUE

p <- partition("REUTERS", places = "kuwait")

#> ... get encoding: latin1

#> ... get cpos and strucs

corpus(p)

#> [1] "REUTERS"

pb <- partition_bundle("REUTERS", s_attribute = "id")

#> ... getting matrix with regions for s-attribute:  id

#> ... generating the partitions

corpus(pb)

#> [1] "REUTERS"