Dispersion of a query or multiple queries

The function returns the frequencies of a query or a multiple queries in sub-partitions defined by one or two dimensions. This is a wrapper function, so the output will depend on the number of queries and dimensions provided.

dispersion(.Object, ...)

# S4 method for partition
dispersion(.Object, query, sAttribute, cqp = FALSE,
  pAttribute = getOption("polmineR.pAttribute"), freq = FALSE, mc = FALSE,
  progress = TRUE, verbose = FALSE)

# S4 method for character
dispersion(.Object, query, sAttribute, cqp = is.cqp,
  pAttribute = getOption("polmineR.pAttribute"), freq = FALSE, mc = FALSE,
  progress = TRUE, verbose = TRUE)

# S4 method for hits
dispersion(.Object, sAttribute, freq = FALSE,
  verbose = TRUE)

Arguments

.Object
a partition object

...
further parameters

query
a character vector containing one or multiple queries

sAttribute
a character vector of length 1 or 2 providing the sAttributes

cqp
if logical, whether the query is a CQP query (TRUE/FALSE), if it is a function that is passed in, the function will be applied to the query to guess whether query is a CQP query

pAttribute
the p-attribute that will be looked up, typically 'word' or 'lemma'

freq
logical, whether to calculate normalized frequencies

mc
logical, whether to use multicore

progress
logical, whether to shop progress

verbose
logical, whether to be verbose

Value

depends on the input, as this is a wrapper function

See also

crosstab-class

count

Examples

# NOT RUN {
  use("polmineR.sampleCorpus")
  test <- partition("PLPRBTTXT", text_year = "2009", pAttribute = NULL)
  integration <- dispersion(
    test, query = "Integration",
    pAttribute = "word", sAttribute = "text_date"
    )
  integration <- dispersion(
    test, "Integration",
    sAttribute = c("text_date", "text_party")
    )
  integration <- dispersion(
    test, '"Integration.*"',
    sAttribute = "text_year", cqp = TRUE
    )
# }