Feature selection by comparison.

The features-method returns a features-object. Several features-objects can be combined into a features_bundle-object.

# S4 method for features
summary(object)

# S4 method for features
show(object)

# S4 method for features_bundle
summary(object)

# S4 method for features
format(x, digits = 2L)

# S4 method for features
view(.Object)

Arguments

object
A features or features_bundle object.

x
A features object.

digits
Integer indicating the number of decimal places (round) or significant digits (signif) to be used.

.Object
a features object.

Details

A set of features objects can be combined into a features_bundle. Typically, a features_bundle will result from applying the features-method on a partition_bundle. See the documentation for bundle to learn about the methods for bundle objects that are available for a features_bundle.

Slots

corpus

The CWB corpus the features are derived from, a character vector of length 1.

p_attribute

Object of class character.

encoding

Object of class character.

corpus

Object of class character.

stat

Object of class data.frame.

size_coi

Object of class integer.

size_ref

Object of class integer.

included

Object of class logical whether corpus of interest is included in reference corpus

method

Object of class character statisticalTest used

call

Object of class character the call that generated the object