Generate a bundle of partitions

A partitionBundle object is a S4 class object. partitionBundle,character-method will create a bundle of partitions, but not yet enriched.

partitionBundle(.Object, ...)

# S4 method for partition
partitionBundle(.Object, sAttribute, values = NULL,
  prefix = "", mc = getOption("polmineR.mc"), verbose = TRUE,
  progress = FALSE, ...)

# S4 method for character
partitionBundle(.Object, sAttribute, values = NULL,
  prefix = "", mc = getOption("polmineR.mc"), verbose = TRUE,
  progress = FALSE, xml = "flat", ...)

# S4 method for context
partitionBundle(.Object, mc = getOption("polmineR.mc"),
  verbose = FALSE, progress = TRUE)

Arguments

.Object
character string, a partition, or a list

...
parameters to be passed into partition-method (see respective documentation)

sAttribute
the s-attribute to vary

values
values the s-attribute provided shall assume

prefix
a character vector that will be attached as a prefix to partition names

mc
logical, whether to use multicore parallelization

verbose
logical, whether to provide progress information

progress
logical, whether to show progress bar

xml
logical

Value

S4 class 'partitionBundle', with list of partition objects in slot 'objects'

See also

partition and bundle-class

Examples

# NOT RUN {
  use("polmineR.sampleCorpus")
  bt2009 <- partition("PLPRBTTXT", text_year = "2009")
  pBundle <- partitionBundle(bt2009, sAttribute = "text_date", progress = TRUE, pAttribute = "word")
  dtm <- as.DocumentTermMatrix(pBundle, col = "count")
  summary(pBundle)
  btBundle <- partitionBundle("PLPRBTTXT", sAttribute = "text_date")
# }