Reset registry directory.

A utility function to reset the environment variable CORPUS_REGISTRY. That may be necessary if you want use a CWB corpus that is not stored in the usual place. In particular, resetting the environment variable is required if you want to use a corpus delivered in a R package,

registry_reset(registryDir = registry(), verbose = TRUE)

Arguments

registryDir
path to the registry directory to be used

verbose
logical, whether to be verbose

Value

the registry directory used before resetting CORPUS_REGISTRY

Details

Resetting the CORPUS_REGISTRY environment variable is also necessary for the interface to CWB corpora.

To get the path to a package that contains a CWB corpus, use system.file (see examples).

See also

Examples

if (FALSE) {
x <- system.file(package = "polmineR", "extdata", "cwb", "registry")
registry_reset(registryDir = x)
}