Get Number of Tokens.

The method will get the number of tokens in a corpus or partition, or the dispersion across one or more s-attributes.

size(x, ...)

# S4 method for character
size(x, s_attribute = NULL, verbose = TRUE, ...)

# S4 method for partition
size(x, s_attribute = NULL, ...)

# S4 method for DocumentTermMatrix
size(x)

# S4 method for TermDocumentMatrix
size(x)

Arguments

x
object to get size(s) for

...
further arguments

s_attribute
character vector with s-attributes (one or more)

verbose
logical, whether to print messages

Value

an integer vector if s_attribute is NULL, a data.table otherweise

Details

One or more s-attributes can be provided to get the dispersion of tokens across one or more dimensions. Two or more s-attributes can lead to reasonable results only if the corpus XML is flat.

See also

Examples

use("polmineR")

#> ... activating corpus: germaparlmini

#> ... activating corpus: reuters

size("GERMAPARLMINI")

#> [1] -11

size("GERMAPARLMINI", s_attribute = "date")

#> Error in check_strucs(corpus = corpus, s_attribute = s_attribute, strucs = struc,     registry = registry): all values of vector strucs need to be >= 0

size("GERMAPARLMINI", s_attribute = c("date", "party"))

#> Error in check_strucs(corpus = corpus, s_attribute = s_attribute, strucs = struc,     registry = registry): all values of vector strucs need to be >= 0

P <- partition("GERMAPARLMINI", date = "2009-11-11")

#> ... get encoding: latin1

#> ... get cpos and strucs

#> Error in rep(0L, times = maxAttr): ungültiges 'times' Argument

size(P, s_attribute = "speaker")

#> Error in size(P, s_attribute = "speaker"): Objekt 'P' nicht gefunden

size(P, s_attribute = "party")

#> Error in size(P, s_attribute = "party"): Objekt 'P' nicht gefunden

size(P, s_attribute = c("speaker", "party"))

#> Error in size(P, s_attribute = c("speaker", "party")): Objekt 'P' nicht gefunden