S4 textstat superclass.

The textstat-class (technically an S4 class) serves as a superclass for the classes features, context, and partition. Usually, the class will not be used directly. It offers a set of standard generic methods (such as head, tail, dim, nrow, colnames) its childs inherit. The core feature of textstat and its childs is a data.table in the slot stat for keeping data on text statistics of a corpus, or a partition.

# S4 method for textstat
name(x)

# S4 method for character
name(x)

# S4 method for textstat,character
name(x) <- value

# S4 method for textstat
round(x, digits = 2L)

# S4 method for textstat
sort(x, by, decreasing = TRUE)

as.bundle(object, ...)

# S4 method for textstat,textstat
+(e1, e2)

# S4 method for textstat
subset(x, ...)

# S3 method for textstat
as.data.table(x)

# S4 method for textstat
show(object)

# S4 method for textstat
p_attributes(.Object)

# S4 method for textstat
knit_print(x, options = knitr::opts_chunk, ...)

# S4 method for textstat
format(x, digits = 2L)

# S4 method for textstat
view(.Object)

Arguments

x
A textstat object.

value
A character vector to assign as name to slot name of a textstat class object.

digits
Number of digits.

by
Column that will serve as the key for sorting.

decreasing
Logical, whether to return decreasing order.

object
a textstat object

...
Further arguments.

e1
A texstat object.

e2
Another texstat object.

.Object
A textstat object.

options
Chunk options.

Details

A head-method will return the first rows of the data.table in the stat-slot. Use argument n to specify the number of rows.

A tail-method will return the last rows of the data.table in the stat-slot. Use argument n to specify the number of rows.

The methods dim, nrow and ncol will return information on the dimensions, the number of rows, or the number of columns of the data.table in the stat-slot, respectively.

Objects derived from the textstat class can be indexed with simple square brackets ("[") to get rows specified by an numeric/integer vector, and with double square brackets ("[[") to get specific columns from the data.table in the slot stat.

The colnames-method will return the column names of the data-table in the slot stat.

The methods as.data.table, and as.data.frame will extract the data.table in the slot stat as a data.table, or data.frame, respectively.

textstat objects can have a name, which can be retrieved, and set using the name-method and name<-, respectively.

The round()-method looks up all numeric columns in the data.table in the stat-slot of the textstat object and rounds values of these columns to the number of decimal places specified by argument digits.

The format()-method returns a pretty-printed and minimized version of the data.table in the stat-slot of the textstat-object: It will round all numeric columns to the number of decimal numbers specified by digits, and drop all columns with token ids. The return value is a data.table.

Slots

p_attribute

Object of class character, p-attribute of the query.

corpus

A corpus specified by a length-one character vector.

stat

A data.table with statistical information.

name

The name of the object.

encoding

A length-one character vector, the encoding of the corpus.

Examples

use("polmineR")

#> ... activating corpus: GERMAPARLMINI

#> ... activating corpus: REUTERS

P <- partition("GERMAPARLMINI", date = ".*", p_attribute = "word", regex = TRUE)

#> ... get encoding: latin1

#> ... get cpos and strucs

#> ... getting counts for p-attribute(s): word

y <- cooccurrences(P, query = "Arbeit")

# Standard generic methods known from data.frames work for objects inheriting
# from the textstat class

head(y)

#>                       word word_id count_partition count_coi count_ref
#> 1:  Arbeitnehmervertretern   13321               1         1         0
#> 2: Arbeitsmarktpolitik.Wir   14664               1         1         0
#> 3:           Bundesanstalt    2471               1         1         0
#> 4:          CSU-Kommission   17449               1         1         0
#> 5:            Erwerbsphase   13353               1         1         0
#> 6:              Exorzisten   13207               1         1         0
#>        exp_coi   exp_ref ll rank_ll
#> 1: 0.007160807 0.9928392 NA     506
#> 2: 0.007160807 0.9928392 NA     512
#> 3: 0.007160807 0.9928392 NA     487
#> 4: 0.007160807 0.9928392 NA     523
#> 5: 0.007160807 0.9928392 NA     507
#> 6: 0.007160807 0.9928392 NA     504

tail(y)

#>     word word_id count_partition count_coi count_ref   exp_coi   exp_ref
#> 1:     )      71            2116         6      2110 15.152268 2100.8477
#> 2:  dass      38            1678         4      1674 12.015835 1665.9842
#> 3:   FDP     290             999         1       998  7.153647  991.8464
#> 4:     ]     829            1160         1      1159  8.306537 1151.6935
#> 5: nicht     324            1814         3      1811 12.989705 1801.0103
#> 6:     /     288            1698         2      1696 12.159051 1685.8409
#>            ll rank_ll
#> 1:  -7.281131     475
#> 2:  -7.311672     476
#> 3:  -8.434214     477
#> 4: -10.459188     478
#> 5: -11.305052     479
#> 6: -13.225234     480

nrow(y)

#> [1] 523

ncol(y)

#> [1] 9

dim(y)

#> [1] 523   9

colnames(y)

#> [1] "word"            "word_id"         "count_partition" "count_coi"      
#> [5] "count_ref"       "exp_coi"         "exp_ref"         "ll"             
#> [9] "rank_ll"        

# Use brackets for indexing 

# NOT RUN {
y[1:25]
y[,c("word", "ll")]
y[1:25, "word"]
y[1:25][["word"]]
y[which(y[["word"]] %in% c("Arbeit", "Sozial"))]
y[ y[["word"]] %in% c("Arbeit", "Sozial") ]
# }