apply a function over a list or bundle

Very similar to lapply, but applicable to bundle-objects, in particular. The purpose of the method is to supply a uniform und convenient parallel backend for the polmineR package. In particular, progress bars are supported (the naming of the method is derived from bla bla).

blapply(x, ...)

# S4 method for list
blapply(x, f, mc = TRUE, progress = TRUE,
  verbose = FALSE, ...)

# S4 method for vector
blapply(x, f, mc = FALSE, progress = TRUE,
  verbose = FALSE, ...)

# S4 method for bundle
blapply(x, f, mc = FALSE, progress = TRUE,
  verbose = FALSE, ...)

Arguments

x
a list or a bundle object

...
further parameters

f
a function that can be applied to each object contained in the bundle, note that it should swallow the parameters mc, verbose and progress (use ... to catch these params )

mc
logical, whether to use multicore - if TRUE, the number of cores will be taken from the polmineR-options

progress
logical, whether to display progress bar

verbose
logical, whether to print intermediate messages

Examples

use("polmineR")

#> ... activating corpus: GERMAPARLMINI

#> ... activating corpus: REUTERS

bt <- partition("GERMAPARLMINI", date = ".*", regex=TRUE)

#> ... get encoding: latin1

#> ... get cpos and strucs

speeches <- as.speeches(bt, s_attribute_date = "date", s_attribute_name = "speaker")

#> ... generating partitions by date

#> ... getting values for s-attribute  date

#> ... number of partitions to be generated:  5

#> ... generating speeches

#> ... generating names

#> ... reordering partitions

#> ... coercing partitions to plpr_partitions

foo <- blapply(speeches, function(x, ...) slot(x, "cpos"))