Get corpus/corpora available or used.

Calling corpus() will return a data.frame listing the corpora described in the active registry directory, and some basic information on the corpora. If object is an object inheriting from the textstat, or the bundle class, the corpus used to generate the object is returned.

corpus(object)

# S4 method for textstat
corpus(object)

# S4 method for kwic
corpus(object)

# S4 method for character
corpus(object)

# S4 method for bundle
corpus(object)

# S4 method for missing
corpus()

Arguments

object
An object inheriting from the textstat or bundle superclasses.

Examples

use("polmineR")

#> ... activating corpus: GERMAPARLMINI

#> ... activating corpus: REUTERS

corpus()

#>          corpus   size template
#> 1 GERMAPARLMINI 222201     TRUE
#> 2       REUTERS   4050     TRUE

p <- partition("REUTERS", places = "kuwait")

#> ... get encoding: latin1

#> ... get cpos and strucs

corpus(p)

#> [1] "REUTERS"

pb <- partition_bundle("REUTERS", s_attribute = "id")

#> ... getting matrix with regions for s-attribute:  id

#> ... generating the partitions

corpus(pb)

#> [1] "REUTERS"