S4 class to wrap information on CWB corpora.

S4 class to wrap information on CWB corpora.

# S4 method for corpus
$(x, name)

# S4 method for corpus,ANY
==(e1, e2)

# S4 method for corpus,ANY
!=(e1, e2)

# S4 method for corpus
%in%(x, table)

# S4 method for corpus
zoom(x, ...)

Arguments

x
A corpus object.

name
An s-attribute that will be assigned as key to a partition.

e1
A first expression, a partition object in this case.

e2
A second expression, the value of an s-attribute.

table
Values a s-attribute shall assume.

...
Further arguments.

Details

The `$`-method will assign the argument name to the slot key and return the modified object.

Slots

corpus

A length-one character vector, a CWB corpus.

data_dir

The directory where the files for the indexed corpus are.

type

The type of the corpus (e.g. "plpr" for a corpus of plenary protocols).

encoding

The encoding of the corpus, given as a length-one character vector.

key

A length-one character vector stating a s-attribute that serves as a key.

See also

Examples

corp <- corpus("GERMAPARLMINI")
corp2 <- corp$speaker
corp2@key

#> [1] "speaker"

x <- corpus("GERMAPARLMINI")
x$date == "2009-10-28"

#> ... get encoding: latin1

#> ... get cpos and strucs

#> ** partition object **
#> corpus:              GERMAPARLMINI 
#> name:                 
#> s-attributes:        date = 2009-10-28 
#> cpos:                81 pairs of corpus positions
#> size:                2793 tokens
#> count:               not available

x <- corpus("GERMAPARLMINI")
x$party != "NA"

#> ... get encoding: latin1

#> ... get cpos and strucs

#> ** partition object **
#> corpus:              GERMAPARLMINI 
#> name:                 
#> s-attributes:        party = CDU_CSU/SPD/FDP/DIE_LINKE/B90_DIE_GRUENEN 
#> cpos:                4046 pairs of corpus positions
#> size:                190891 tokens
#> count:               not available

x <- corpus("GERMAPARLMINI")
x$date %in% c("2009-10-27", "2009-10-28")

#> ... get encoding: latin1

#> ... get cpos and strucs

#> ** partition object **
#> corpus:              GERMAPARLMINI 
#> name:                 
#> s-attributes:        date = 2009-10-27/2009-10-28 
#> cpos:                227 pairs of corpus positions
#> size:                12134 tokens
#> count:               not available

x <- corpus("GERMAPARLMINI")
y <- zoom(x, date == "2009-10-28")

#> ... get encoding: latin1

#> ... get cpos and strucs

x <- partition("GERMAPARLMINI", interjection = "speech")

#> ... get encoding: latin1

#> ... get cpos and strucs

m <- zoom(x, date == "2009-10-28" & speaker == "Angela Dorothea Merkel")

#> ... Setting up partition 

#> ... getting cpos and strucs

#> ... Setting up partition 

#> ... getting cpos and strucs

not_unknown <- zoom(x, party != c("NA", "FDP"))

#> ... Setting up partition 

#> ... getting cpos and strucs

s_attributes(not_unknown, "party")

#> [1] "CDU_CSU"         "SPD"             "DIE_LINKE"       "B90_DIE_GRUENEN"