Decode corpus or subcorpus.

Decode corpus or subcorpus and return class specified by argument to.

decode(.Object, ...)

# S4 method for character
decode(.Object, to = c("data.table", "Annotation"),
  ...)

# S4 method for slice
decode(.Object, to = "data.table")

# S4 method for partition
decode(.Object, to = "data.table")

Arguments

.Object
The corpus or subcorpus to decode.

...
Further arguments.

to
The class of the returned object, stated as a length-one character vector.

Value

The return value will correspond to the class specified by argument to.

Details

The primary purpose of the method is type conversion. By obtaining the corpus or subcorpus in the format specified by the argument to, the data can be processed with tools that do not rely on the Corpus Workbench (CWB). Supported output formats are data.table (which can be converted to a data.frame or tibble easily) or an Annotation object as defined in the package NLP. Another purpose of decoding the corpus can be to rework it, and to re-import it into the CWB (e.g. using the cwbtools-package).

An earlier version of the method included an option to decode a single s-attribute, which is not supported any more. See the s_attribute_decode function of the package RcppCWB.

See also

Examples

use("polmineR")

#> ... activating corpus: GERMAPARLMINI

#> ... activating corpus: REUTERS

# Decode corpus as data.table
dt <- decode("GERMAPARLMINI", to = "data.table")

#> decoding p-attribute:word

#> decoding p-attribute:pos

#> decoding s-attribute:interjection

#> decoding s-attribute:date

#> decoding s-attribute:party

#> decoding s-attribute:speaker

#> assembling data.table

# Decode a subcorpus
sc <- subset(corpus("GERMAPARLMINI"), speaker == "Angela Dorothea Merkel")
dt <- decode(sc, to = "data.table")

#> ... decoding p_attribute word

#> ... decoding p_attribute pos

# Decode partition
P <- partition("REUTERS", places = "kuwait", regex = TRUE)

#> ... get encoding: latin1

#> ... get cpos and strucs

dt <- decode(P)

#> ... decoding p_attribute word

# Previous versions of polmineR offered an option to decode a single
# s-attribute. This is how you could proceed to get a table with metadata.
dt[, "word" := NULL]

#>      cpos struc  id
#>   1:  753     5 236
#>   2:  754     5 236
#>   3:  755     5 236
#>   4:  756     5 236
#>   5:  757     5 236
#>  ---               
#> 656: 3169    13 353
#> 657: 3170    13 353
#> 658: 3171    13 353
#> 659: 3172    13 353
#> 660: 3173    13 353
#>                                                                                                                                                topics_cat
#>   1: crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|c("crude", "ship")|crude|crude|crude|crude|c("crude", "nat-gas")
#>   2: crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|c("crude", "ship")|crude|crude|crude|crude|c("crude", "nat-gas")
#>   3: crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|c("crude", "ship")|crude|crude|crude|crude|c("crude", "nat-gas")
#>   4: crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|c("crude", "ship")|crude|crude|crude|crude|c("crude", "nat-gas")
#>   5: crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|c("crude", "ship")|crude|crude|crude|crude|c("crude", "nat-gas")
#>  ---                                                                                                                                                     
#> 656: crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|c("crude", "ship")|crude|crude|crude|crude|c("crude", "nat-gas")
#> 657: crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|c("crude", "ship")|crude|crude|crude|crude|c("crude", "nat-gas")
#> 658: crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|c("crude", "ship")|crude|crude|crude|crude|c("crude", "nat-gas")
#> 659: crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|c("crude", "ship")|crude|crude|crude|crude|c("crude", "nat-gas")
#> 660: crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|crude|c("crude", "ship")|crude|crude|crude|crude|c("crude", "nat-gas")
#>              places language
#>   1: kuwait|ecuador       en
#>   2: kuwait|ecuador       en
#>   3: kuwait|ecuador       en
#>   4: kuwait|ecuador       en
#>   5: kuwait|ecuador       en
#>  ---                        
#> 656:         kuwait       en
#> 657:         kuwait       en
#> 658:         kuwait       en
#> 659:         kuwait       en
#> 660:         kuwait       en

dt[,
  {list(cpos_left = min(.SD[["cpos"]]), cpos_right = max(.SD[["cpos"]]), id = unique(.SD[["id"]]))},
  by = "struc"
  ]

#>    struc cpos_left cpos_right  id
#> 1:     5       753       1217 236
#> 2:    11      2874       2965 349
#> 3:    13      3071       3173 353

# Decode subcorpus as Annotation object
if (FALSE) {
if (requireNamespace("NLP")){
  library(NLP)
  p <- subset(corpus("GERMAPARLMINI"), date == "2009-11-10" & speaker == "Angela Dorothea Merkel")
  s <- as(p, "String")
  a <- as(p, "Annotation")

  # The beauty of having this NLP Annotation object is that you can now use 
  # the different annotators of the openNLP package. Here, just a short scenario
  # how you can have a look at the tokenized words and the sentences.

  words <- s[a[a$type == "word"]]
  sentences <- s[a[a$type == "sentence"]] # does not yet work perfectly for plenary protocols 
}
}