Dispersion of a query or multiple queries.

The method returns counts (optionally frequencies) of a query or a multiple queries in subcorpora defined by one or two s-attributes.

dispersion(.Object, ...)

# S4 method for slice
dispersion(
  .Object,
  query,
  s_attribute,
  cqp = FALSE,
  p_attribute = getOption("polmineR.p_attribute"),
  freq = FALSE,
  mc = FALSE,
  progress = FALSE,
  verbose = FALSE,
  ...
)

# S4 method for partition
dispersion(
  .Object,
  query,
  s_attribute,
  cqp = FALSE,
  p_attribute = getOption("polmineR.p_attribute"),
  freq = FALSE,
  mc = FALSE,
  progress = TRUE,
  verbose = FALSE,
  ...
)

# S4 method for subcorpus
dispersion(
  .Object,
  query,
  s_attribute,
  cqp = FALSE,
  p_attribute = getOption("polmineR.p_attribute"),
  freq = FALSE,
  mc = FALSE,
  progress = TRUE,
  verbose = FALSE,
  ...
)

# S4 method for corpus
dispersion(
  .Object,
  query,
  s_attribute,
  cqp = is.cqp,
  p_attribute = getOption("polmineR.p_attribute"),
  freq = FALSE,
  mc = FALSE,
  progress = FALSE,
  verbose = FALSE,
  ...
)

# S4 method for character
dispersion(
  .Object,
  query,
  s_attribute,
  cqp = is.cqp,
  p_attribute = getOption("polmineR.p_attribute"),
  freq = FALSE,
  mc = FALSE,
  progress = TRUE,
  verbose = TRUE,
  ...
)

# S4 method for hits
dispersion(.Object, source, s_attribute, freq = FALSE, verbose = TRUE, ...)

# S4 method for remote_corpus
dispersion(.Object, ...)

# S4 method for remote_subcorpus
dispersion(.Object, ...)

Arguments

.Object
A partition object or a corpus provided by a character string.

...
Further parameters.

query
A character vector stating one or multiple queries.

s_attribute
A character vector (length 1 or 2) providing s-attributes.

cqp
If logical, whether the query is a CQP query, if it is a function that is passed in, the function will be applied to the query to guess whether query is a CQP query

p_attribute
Length one character vector, the p-attribute that will be looked up (typically 'word' or 'lemma').

freq
A logical value, whether to calculate normalized frequencies.

mc
A logical value, whether to use multicore.

progress
A logical value, whether to show progress.

verbose
A logical value, whether to be verbose.

source
The source of the evaluation the hits reported in .Object are based on, a corpus, subcorpus or partition object.

Value

depends on the input, as this is a wrapper function

A data.table.

See also

Examples

use("polmineR")

#> ... activating corpus: GERMAPARLMINI (version: 0.0.1 | build date: 2019-02-23)

#> ... activating corpus: REUTERS

dispersion("GERMAPARLMINI", query = "Integration", s_attribute = "date")

#>          query       date count
#> 1: Integration 2009-10-27     1
#> 2: Integration 2009-10-28     0
#> 3: Integration 2009-11-10     7
#> 4: Integration 2009-11-11    15
#> 5: Integration 2009-11-12     0

test <- partition("GERMAPARLMINI", date = ".*", p_attribute = NULL, regex = TRUE)

#> ... get encoding: latin1

#> ... get cpos and strucs

integration <- dispersion(
  test, query = "Integration",
  p_attribute = "word", s_attribute = "date"
)
integration <- dispersion(test, "Integration", s_attribute = c("date", "party"))
integration <- dispersion(test, '"Integration.*"', s_attribute = "date", cqp = TRUE)