Get corpus/partition type.

To generate fulltext output, different templates can be used with a behavior that depends on the type of a corpus. get_type will return the type of corpus if it is a specialized one, or NULL.

get_type(.Object)

# S4 method for corpus
get_type(.Object)

# S4 method for character
get_type(.Object)

# S4 method for partition
get_type(.Object)

# S4 method for subcorpus
get_type(.Object)

# S4 method for partition_bundle
get_type(.Object)

# S4 method for subcorpus_bundle
get_type(.Object)

Arguments

.Object
A partition, partition_bundle, Corpus object, or a length-one character vector indicating a CWB corpus.

Details

When generating a partition, the corpus type will be prefixed to the class that is generated (separated by underscore). If the corpus type is not NULL, a class inheriting from the partition-class is instantiated. Note that at this time, only plpr_partition and press_partition is implemented.

Examples

use("polmineR")

#> ... activating corpus: GERMAPARLMINI (version: 0.0.1 | build date: 2019-02-23)

#> ... activating corpus: REUTERS

get_type("GERMAPARLMINI")

#> [1] "plpr"

p <- partition("GERMAPARLMINI", date = "2009-10-28")

#> ... get encoding: latin1

#> ... get cpos and strucs

get_type(p)

#> [1] "plpr"

is(p)

#> [1] "plpr_partition" "partition"      "count"          "slice"         
#> [5] "textstat"      

pb <- partition_bundle("GERMAPARLMINI", s_attribute = "date")
get_type(pb)

#> [1] "plpr"

get_type("REUTERS") # returns NULL - no specialized corpus

#> NULL