Get Hits.

Get hits for a (set of) queries, optionally with s-attribute values.

hits(.Object, ...)

# S4 method for character
hits(.Object, query, cqp = FALSE, sAttribute = NULL,
  pAttribute = "word", size = FALSE, freq = FALSE, mc = FALSE,
  verbose = TRUE, progress = TRUE)

# S4 method for partition
hits(.Object, query, cqp = FALSE, sAttribute = NULL,
  pAttribute = "word", size = FALSE, freq = FALSE, mc = FALSE,
  progress = FALSE, verbose = TRUE)

# S4 method for partitionBundle
hits(.Object, query,
  pAttribute = getOption("polmineR.pAttribute"), size = TRUE,
  freq = FALSE, mc = getOption("polmineR.mc"), progress = FALSE,
  verbose = TRUE)

# S4 method for hits
sample(x, size)

# S4 method for context
hits(.Object, sAttribute = NULL, verbose = TRUE)

Arguments

.Object
a character, partition or partitionBundle object

...
further parameters

query
a (optionally named, see datails) character vector with one or more queries

cqp
either logical (TRUE if query is a CQP query), or a function to check whether query is a CQP query or not

sAttribute
s-attributes

pAttribute
p-attribute (will be passed into cpos)

size
logical - return size of subcorpus

freq
logcial - return relative frequencies

mc
logical, whether to use multicore

verbose
logical

progress
logical, whether to show progress bar

x
a hits object

Details

If the query character vector is named, the names of the query occurr in the data.table that is returned rather than the queries.

If freq is TRUE, the data.table returned in the DT-slot will deliberately include the subsets of the partition/corpus with no hits (query is NA, count is 0).

Slots

dt

a "data.table"

corpus

a "character" vector

query

Object of class "character"