kwic (S4 class)

S4 class for organizing information for kwic/concordance output. A set of standard generics (show, as.character, as.data.frame, length, sample, subset) as well as indexing is implemented to process kwic class objects (see 'Usage'). See section 'Details' for the enrich, view and knit_print methods.

# S4 method for kwic
show(object)

# S4 method for kwic
knit_print(x, pagelength = getOption("polmineR.pagelength"),
  options = knitr::opts_chunk, ...)

# S4 method for kwic
as.character(x, fmt = "%s")

# S4 method for kwic,ANY,ANY,ANY
[(x, i)

# S4 method for kwic
subset(x, ...)

# S4 method for kwic
as.data.frame(x)

# S4 method for kwic
length(x)

# S4 method for kwic
sample(x, size)

# S4 method for kwic
enrich(.Object, s_attributes = NULL, table = FALSE, ...)

# S4 method for kwic
view(.Object)

Arguments

object
A kwic class object.

x
A kwic class object.

pagelength
The number of kwic lines displayed per page in the datatables htmlwidget that is returned.

options
Chunk options.

...
Used for backwards compatibility.

fmt
A format string passed into sprintf to format the node of a KWIC display.

i
Single integer value, the kwic line for which the fulltext shall be inspected.

size
An integer, subset size for sampling.

.Object
A kwic class object.

s_attributes
Character vector of s-attributes with metainformation.

table
Logical, whether to turn cpos data.table into data.frame for output.

Details

The knit_print will be called by knitr when processing code chunks in Rmarkdown documents to include a htmlwidget into the resulting html document. It may be necessary to explicitly state "render=knit_print" in the chunk options.

The subset-method will apply subset to the table in the slot table, for filtering query results based on metadata (i.e. s-attributes) that need to be present.

The enrich method is used to generate the actual output for the kwic method. If param table is TRUE, corpus positions will be turned into a data.frame with the concordance lines. If param s_attributes is a character vector with s-attributes, the respective s-attributes will be added as columns to the table with concordance lines.

Slots

metadata

A character vector with s-attributes of the metadata that are to be displayed.

left

An integer value, words to the left of the query match.

right

An integer value, words to the right of the query match.

corpus

Length-one character vector, the CWB corpus.

cpos

A data.table with the columns "hit_no", "cpos", "position", "word_id", "word" and "direction".

table

A data.frame, a table with columns "left", "node", "right", and metadata, if the object has been enriched.

encoding

A length-one character vector with the encoding of the corpus.

labels

A character vector with labels.

categories

A character vector.

See also

Examples

use("polmineR")

#> ... activating corpus: germaparlmini

#> ... activating corpus: reuters

K <- kwic("GERMAPARLMINI", "Integration")
length(K)

#> [1] 23

K_min <- K[1]
K_min <- K[1:5]
oil <- kwic("REUTERS", query = "oil")
as.character(oil)

#>  [1] "cut its contract prices for crude <i>oil</i> by 1.50 dlrs a barrel The"                             
#>  [2] "made in the light of falling <i>oil</i> product prices and a weak crude"                            
#>  [3] "product prices and a weak crude <i>oil</i> market a company spokeswoman said Diamond"               
#>  [4] "latest in a line of U.S <i>oil</i> companies that have cut its contract"                            
#>  [5] "the last two days citing weak <i>oil</i> markets Reuter OPEC may be forced"                         
#>  [6] "to halt the current slide in <i>oil</i> prices oil industry analysts said The"                      
#>  [7] "the current slide in oil prices <i>oil</i> industry analysts said The movement to"                  
#>  [8] "analysts said The movement to higher <i>oil</i> prices was never to be as"                          
#>  [9] "Energy Research Associates CERA Analysts and <i>oil</i> industry sources said the problem OPEC"     
#> [10] "the problem OPEC faces is excess <i>oil</i> supply in world oil markets OPEC's"                     
#> [11] "is excess oil supply in world <i>oil</i> markets OPEC's problem is not a"                           
#> [12] "in that way said Paul Mlotok <i>oil</i> analyst with Salomon Brothers Inc He"                       
#> [13] "wishes to regain the initiative in <i>oil</i> prices But some other analysts were"                  
#> [14] "manager for one of the major <i>oil</i> companies who spoke on condition that"                      
#> [15] "the winter demand to sell their <i>oil</i> but in late March and April"                             
#> [16] "said that the demand for OPEC <i>oil</i> has been rising through the first"                         
#> [17] "its production Demand for their OPEC <i>oil</i> is clearly above 15.8 mln bpd"                      
#> [18] "price it will pay for crude <i>oil</i> 64 Canadian cts a barrel effective"                          
#> [19] "Texaco Canada last changed its crude <i>oil</i> postings on Feb 19 Reuter Marathon"                 
#> [20] "pay for all grades of crude <i>oil</i> one dlr a barrel effective today"                            
#> [21] "an emergency OPEC meeting to review <i>oil</i> policies after recent weakness in world"             
#> [22] "policies after recent weakness in world <i>oil</i> prices Sheikh Ali al Khalifa al"                 
#> [23] "organisation Traders and analysts in international <i>oil</i> markets estimate OPEC is producing up"
#> [24] "delivered a challenge to any international <i>oil</i> company that declared Kuwait sold below"      
#> [25] "Kuwait had guaranteed markets for its <i>oil</i> because of its local and international"            
#> [26] "pricing committee Referring to pressure by <i>oil</i> companies on OPEC members in apparent"        
#> [27] "a new report To counter falling <i>oil</i> revenues the government has launched a"                  
#> [28] "months to boost exports outside the <i>oil</i> sector and attract new investment Indonesia"         
#> [29] "last year s fall in world <i>oil</i> prices which forced it to devalue"                             
#> [30] "petroleum industry Growth in the non <i>oil</i> sector is low because of weak"                      
#> [31] "in halting the current decline in <i>oil</i> prices Oil industry sources said yesterday"            
#> [32] "Arab producers had had difficulty selling <i>oil</i> at official OPEC prices but Kuwait"            
#> [33] "3.7500 03 yesterday REUTER The Gulf <i>oil</i> state of Qatar recovering slightly from"             
#> [34] "from last year's decline in world <i>oil</i> prices announced its first budget since"               
#> [35] "87 budget due to uncertainty surrounding <i>oil</i> revenues Sheikh Abdul Aziz said that"           
#> [36] "There was also no projection for <i>oil</i> revenue Qatar an OPEC member has"                       
#> [37] "Our expectations of positive signs regarding <i>oil</i> price trends foremost among them OPEC's"    
#> [38] "December's OPEC accord to boost world <i>oil</i> prices and stabilise the market the"               
#> [39] "the recent fall in free market <i>oil</i> prices Nazer said Saudi Arabia is"                        
#> [40] "and it will never sell its <i>oil</i> at prices below the pronounced prices"                        
#> [41] "followed a year of turmoil on <i>oil</i> markets which saw prices slump briefly"                    
#> [42] "accord was shown clearly in the <i>oil</i> market He said contacts among members"                   
#> [43] "prices Traders and analysts in international <i>oil</i> markets estimate OPEC is producing up"      
#> [44] "was over producing REUTER Saudi crude <i>oil</i> output last month fell to an"                      
#> [45] "3.8 mln bpd in January Gulf <i>oil</i> sources said They said exports from"                         
#> [46] "last December to back new official <i>oil</i> prices averaging 18 dlrs a barrel"                    
#> [47] "dlr below Opec levels Saudi Arabian <i>oil</i> minister Hisham Nazer yesterday reiterated the"      
#> [48] "Jubail export refineries They put crude <i>oil</i> exports through Yanbu at 100,000 bpd"            
#> [49] "around 200,000 bpd each REUTER Deputy <i>oil</i> ministers from six Gulf Arab states"               
#> [50] "today to discuss coordination of crude <i>oil</i> marketing the official Emirates news agency"      
#> [51] "Doha by Gulf Cooperation Council GCC <i>oil</i> ministers to help each other market"                
#> [52] "help each other market their crude <i>oil</i> Four of the GCC states Saudi"                         
#> [53] "December's OPEC accord to boost world <i>oil</i> prices and stabilize the market the"               
#> [54] "the recent fall in free market <i>oil</i> prices Nazer said Saudi Arabia is"                        
#> [55] "and it will never sell its <i>oil</i> at prices below the pronounced prices"                        
#> [56] "OPEC agreed to cut its total <i>oil</i> output ceiling by 7.25 pct and"                             
#> [57] "18 dollars a barrel Reuter Kuwait's <i>oil</i> minister said in a newspaper interview"              
#> [58] "after the recent weakness in world <i>oil</i> prices Sheikh Ali al Khalifa al"                      
#> [59] "barrels of crude daily bpd Crude <i>oil</i> prices fell sharply last week as"                       
#> [60] "fell sharply last week as international <i>oil</i> traders and analysts estimated the 13"           
#> [61] "Philadelphia was closed when a Cypriot <i>oil</i> tanker Seapride II ran aground after"             
#> [62] "said He said there was no <i>oil</i> spill but the ship is lodged"                                  
#> [63] "on the high tide After delivering <i>oil</i> to a refinery in Paulsboro New"                        
#> [64] "present and future impact of low <i>oil</i> prices on the domestic oil industry"                    
#> [65] "low oil prices on the domestic <i>oil</i> industry U.S policy now is to"                            
#> [66] "also called for new research for <i>oil</i> exploration and development techniques It predicted"    
#> [67] "instead that such moves as increasing <i>oil</i> reserves and more exploration and development"     
#> [68] "present and future impact of low <i>oil</i> prices on the domestic oil industry"                    
#> [69] "low oil prices on the domestic <i>oil</i> industry U.S policy now is to"                            
#> [70] "also called for new research for <i>oil</i> exploration and development techniques It predicted"    
#> [71] "the full economic benefits of cheap <i>oil</i> But the group did not strongly"                      
#> [72] "instead that such moves as increasing <i>oil</i> reserves and more exploration and development"     
#> [73] "lowered its posted prices for crude <i>oil</i> one to 1.50 dlrs a barrel"                           
#> [74] "for West Coast grades of crude <i>oil</i> the company said Reuter The New"                          
#> [75] "On April one NYMEX will allow <i>oil</i> traders that do not hold a"                                
#> [76] "spokeswoman This will change the way <i>oil</i> is transacted in the real world"                    
#> [77] "will serve the industry because the <i>oil</i> market does not close when NYMEX"                    
#> [78] "a CFTC spokeswoman Reuter Argentine crude <i>oil</i> production was down 10.8 pct in"