kwic (S4 class)

S4 class for organizing information for kwic/concordance output. A set of standard generics (show, as.character, as.data.frame, length, sample, subset) as well as indexing is implemented to process kwic class objects (see 'Usage'). See section 'Details' for the enrich, view and knit_print methods.

# S4 method for kwic
show(object)

# S4 method for kwic
knit_print(x, pagelength = getOption("polmineR.pagelength"),
  options = knitr::opts_chunk, ...)

# S4 method for kwic
as.character(x, fmt = "%s")

# S4 method for kwic,ANY,ANY,ANY
[(x, i)

# S4 method for kwic
subset(x, ...)

# S4 method for kwic
as.data.frame(x)

# S4 method for kwic
length(x)

# S4 method for kwic
sample(x, size)

# S4 method for kwic
enrich(.Object, s_attributes = NULL, table = FALSE, ...)

# S4 method for kwic
view(.Object)

Arguments

object
A kwic class object.

x
A kwic class object.

pagelength
The number of kwic lines displayed per page in the datatables htmlwidget that is returned.

options
Chunk options.

...
Used for backwards compatibility.

fmt
A format string passed into sprintf to format the node of a KWIC display.

i
Single integer value, the kwic line for which the fulltext shall be inspected.

size
An integer, subset size for sampling.

.Object
A kwic class object.

s_attributes
Character vector of s-attributes with metainformation.

table
Logical, whether to turn cpos data.table into data.frame for output.

Details

The knit_print will be called by knitr when processing code chunks in Rmarkdown documents to include a htmlwidget into the resulting html document. It may be necessary to explicitly state "render=knit_print" in the chunk options.

The subset-method will apply subset to the table in the slot table, for filtering query results based on metadata (i.e. s-attributes) that need to be present.

The enrich method is used to generate the actual output for the kwic method. If param table is TRUE, corpus positions will be turned into a data.frame with the concordance lines. If param s_attributes is a character vector with s-attributes, the respective s-attributes will be added as columns to the table with concordance lines.

Slots

metadata

A character vector with s-attributes of the metadata that are to be displayed.

p_attribute

The p-attribute for which the context has been generated.

left

An integer value, words to the left of the query match.

right

An integer value, words to the right of the query match.

corpus

Length-one character vector, the CWB corpus.

cpos

A data.table with the columns "hit_no", "cpos", "position", "word_id", "word" and "direction".

table

A data.frame, a table with columns "left", "node", "right", and metadata, if the object has been enriched.

encoding

A length-one character vector with the encoding of the corpus.

labels

A character vector with labels.

categories

A character vector.

See also

Examples

use("polmineR")

#> ... activating corpus: GERMAPARLMINI

#> ... activating corpus: REUTERS

K <- kwic("GERMAPARLMINI", "Integration")
length(K)

#> [1] 23

K_min <- K[1]
K_min <- K[1:5]
oil <- kwic("REUTERS", query = "oil")
as.character(oil)

#>  [1] "its contract prices for crude <i>oil</i> by 1.50 dlrs a barrel"                       
#>  [2] "in the light of falling <i>oil</i> product prices and a weak"                         
#>  [3] "prices and a weak crude <i>oil</i> market a company spokeswoman said"                 
#>  [4] "in a line of U.S <i>oil</i> companies that have cut its"                              
#>  [5] "last two days citing weak <i>oil</i> markets Reuter OPEC may be"                      
#>  [6] "halt the current slide in <i>oil</i> prices oil industry analysts said"               
#>  [7] "current slide in oil prices <i>oil</i> industry analysts said The movement"           
#>  [8] "said The movement to higher <i>oil</i> prices was never to be"                        
#>  [9] "Research Associates CERA Analysts and <i>oil</i> industry sources said the problem"   
#> [10] "problem OPEC faces is excess <i>oil</i> supply in world oil markets"                  
#> [11] "excess oil supply in world <i>oil</i> markets OPEC's problem is not"                  
#> [12] "that way said Paul Mlotok <i>oil</i> analyst with Salomon Brothers Inc"               
#> [13] "to regain the initiative in <i>oil</i> prices But some other analysts"                
#> [14] "for one of the major <i>oil</i> companies who spoke on condition"                     
#> [15] "winter demand to sell their <i>oil</i> but in late March and"                         
#> [16] "that the demand for OPEC <i>oil</i> has been rising through the"                      
#> [17] "production Demand for their OPEC <i>oil</i> is clearly above 15.8 mln"                
#> [18] "it will pay for crude <i>oil</i> 64 Canadian cts a barrel"                            
#> [19] "Canada last changed its crude <i>oil</i> postings on Feb 19 Reuter"                   
#> [20] "for all grades of crude <i>oil</i> one dlr a barrel effective"                        
#> [21] "emergency OPEC meeting to review <i>oil</i> policies after recent weakness in"        
#> [22] "after recent weakness in world <i>oil</i> prices Sheikh Ali al Khalifa"               
#> [23] "Traders and analysts in international <i>oil</i> markets estimate OPEC is producing"  
#> [24] "a challenge to any international <i>oil</i> company that declared Kuwait sold"        
#> [25] "had guaranteed markets for its <i>oil</i> because of its local and"                   
#> [26] "committee Referring to pressure by <i>oil</i> companies on OPEC members in"           
#> [27] "new report To counter falling <i>oil</i> revenues the government has launched"        
#> [28] "to boost exports outside the <i>oil</i> sector and attract new investment"            
#> [29] "year s fall in world <i>oil</i> prices which forced it to"                            
#> [30] "industry Growth in the non <i>oil</i> sector is low because of"                       
#> [31] "halting the current decline in <i>oil</i> prices Oil industry sources said"           
#> [32] "producers had had difficulty selling <i>oil</i> at official OPEC prices but"          
#> [33] "03 yesterday REUTER The Gulf <i>oil</i> state of Qatar recovering slightly"           
#> [34] "last year's decline in world <i>oil</i> prices announced its first budget"            
#> [35] "budget due to uncertainty surrounding <i>oil</i> revenues Sheikh Abdul Aziz said"     
#> [36] "was also no projection for <i>oil</i> revenue Qatar an OPEC member"                   
#> [37] "expectations of positive signs regarding <i>oil</i> price trends foremost among them" 
#> [38] "OPEC accord to boost world <i>oil</i> prices and stabilise the market"                
#> [39] "recent fall in free market <i>oil</i> prices Nazer said Saudi Arabia"                 
#> [40] "it will never sell its <i>oil</i> at prices below the pronounced"                     
#> [41] "a year of turmoil on <i>oil</i> markets which saw prices slump"                       
#> [42] "was shown clearly in the <i>oil</i> market He said contacts among"                    
#> [43] "Traders and analysts in international <i>oil</i> markets estimate OPEC is producing"  
#> [44] "over producing REUTER Saudi crude <i>oil</i> output last month fell to"               
#> [45] "mln bpd in January Gulf <i>oil</i> sources said They said exports"                    
#> [46] "December to back new official <i>oil</i> prices averaging 18 dlrs a"                  
#> [47] "below Opec levels Saudi Arabian <i>oil</i> minister Hisham Nazer yesterday reiterated"
#> [48] "export refineries They put crude <i>oil</i> exports through Yanbu at 100,000"         
#> [49] "200,000 bpd each REUTER Deputy <i>oil</i> ministers from six Gulf Arab"               
#> [50] "to discuss coordination of crude <i>oil</i> marketing the official Emirates news"     
#> [51] "by Gulf Cooperation Council GCC <i>oil</i> ministers to help each other"              
#> [52] "each other market their crude <i>oil</i> Four of the GCC states"                      
#> [53] "OPEC accord to boost world <i>oil</i> prices and stabilize the market"                
#> [54] "recent fall in free market <i>oil</i> prices Nazer said Saudi Arabia"                 
#> [55] "it will never sell its <i>oil</i> at prices below the pronounced"                     
#> [56] "agreed to cut its total <i>oil</i> output ceiling by 7.25 pct"                        
#> [57] "dollars a barrel Reuter Kuwait's <i>oil</i> minister said in a newspaper"             
#> [58] "the recent weakness in world <i>oil</i> prices Sheikh Ali al Khalifa"                 
#> [59] "of crude daily bpd Crude <i>oil</i> prices fell sharply last week"                    
#> [60] "sharply last week as international <i>oil</i> traders and analysts estimated the"     
#> [61] "was closed when a Cypriot <i>oil</i> tanker Seapride II ran aground"                  
#> [62] "He said there was no <i>oil</i> spill but the ship is"                                
#> [63] "the high tide After delivering <i>oil</i> to a refinery in Paulsboro"                 
#> [64] "and future impact of low <i>oil</i> prices on the domestic oil"                       
#> [65] "oil prices on the domestic <i>oil</i> industry U.S policy now is"                     
#> [66] "called for new research for <i>oil</i> exploration and development techniques It"     
#> [67] "that such moves as increasing <i>oil</i> reserves and more exploration and"           
#> [68] "and future impact of low <i>oil</i> prices on the domestic oil"                       
#> [69] "oil prices on the domestic <i>oil</i> industry U.S policy now is"                     
#> [70] "called for new research for <i>oil</i> exploration and development techniques It"     
#> [71] "full economic benefits of cheap <i>oil</i> But the group did not"                     
#> [72] "that such moves as increasing <i>oil</i> reserves and more exploration and"           
#> [73] "its posted prices for crude <i>oil</i> one to 1.50 dlrs a"                            
#> [74] "West Coast grades of crude <i>oil</i> the company said Reuter The"                    
#> [75] "April one NYMEX will allow <i>oil</i> traders that do not hold"                       
#> [76] "This will change the way <i>oil</i> is transacted in the real"                        
#> [77] "serve the industry because the <i>oil</i> market does not close when"                 
#> [78] "CFTC spokeswoman Reuter Argentine crude <i>oil</i> production was down 10.8 pct"