Initialize a partition.

Create a subcorpus and keep it in an object of the partition class. If defined, counts are performed for the p-attribute defined by the parameter p_attribute.

partition(.Object, ...)

# S4 method for character
partition(.Object, def = NULL, name = "",
  encoding = NULL, p_attribute = NULL, regex = FALSE, xml = "flat",
  decode = TRUE, type = get_type(.Object), mc = FALSE, verbose = TRUE,
  ...)

# S4 method for environment
partition(.Object, slots = c("name", "corpus", "size",
  "p_attribute"))

# S4 method for partition
partition(.Object, def = NULL, name = "",
  regex = FALSE, p_attribute = NULL, decode = TRUE, xml = NULL,
  verbose = TRUE, mc = FALSE, ...)

# S4 method for Corpus
partition(.Object, def = NULL, name = "",
  encoding = NULL, regex = FALSE, xml = "flat",
  type = get_type(.Object), verbose = TRUE, ...)

Arguments

.Object
A length-one character-vector, the CWB corpus to be used.

...
Arguments to define partition (see examples).

def
A named list of character vectors of s-attribute values, the names are the s-attributes (see details and examples)

name
A name for the new partition object, defaults to "".

encoding
The encoding of the corpus (typically "LATIN1 or "(UTF-8)), if NULL, the encoding provided in the registry file of the corpus (charset="...") will be used.

p_attribute
The p-attribute(s) for which a count is performed.

regex
A logical value (defaults to FALSE).

xml
Either 'flat' (default) or 'nested'.

decode
Logical, whether to turn token ids to strings (set FALSE to minimize object size / memory consumption) in data.table with counts.

type
A length-one character vector specifying the type of corpus / partition (e.g. "plpr")

mc
Whether to use multicore (for counting terms).

verbose
Logical, whether to be verbose.

Value

An object of the S4 class partition.

Details

If .Object is a length-one character vector, a subcorpus/partition for the corpus defined be .Object is generated.

If .Object is an environment (typically .GlobalEnv), the partition objects present in the environment are listed.

If .Object is a partition object, a subcorpus of the subcorpus is generated.

If .Object is a Corpus object, preparing the partition may work more efficiently than if .Object is a length-one character vector.

See also

Examples

use("polmineR")

#> ... activating corpus: germaparlmini

#> ... activating corpus: reuters

spd <- partition("GERMAPARLMINI", party = "SPD", interjection = "speech")

#> ... get encoding: latin1

#> ... get cpos and strucs

kauder <- partition("GERMAPARLMINI", speaker = "Volker Kauder", p_attribute = "word")

#> ... get encoding: latin1

#> ... get cpos and strucs

#> ... getting counts for p-attribute(s): word

merkel <- partition("GERMAPARLMINI", speaker = ".*Merkel", p_attribute = "word", regex = TRUE)

#> ... get encoding: latin1

#> ... get cpos and strucs

#> ... getting counts for p-attribute(s): word

s_attributes(merkel, "date")

#> [1] "2009-10-28" "2009-11-10"

s_attributes(merkel, "speaker")

#> [1] "Angela Dorothea Merkel"

merkel <- partition(
  "GERMAPARLMINI", speaker = "Angela Dorothea Merkel",
  date = "2009-11-10", interjection = "speech", p_attribute = "word"
  )

#> ... get encoding: latin1

#> ... get cpos and strucs

#> ... getting counts for p-attribute(s): word

merkel <- subset(merkel, !word %in% punctuation)
merkel <- subset(merkel, !word %in% tm::stopwords("de"))

# a certain defined time segment
days <- seq(
  from = as.Date("2009-10-28"),
  to = as.Date("2009-11-11"),
  by = "1 day"
)
period <- partition("GERMAPARLMINI", date = days)

#> ... get encoding: latin1

#> ... get cpos and strucs