Partition class and methods.

The partition class is used to manage subcorpora. It is an S4 class, and a set of methods is defined for the class. The class inherits from the classes count and textstat.

# S4 method for partition
summary(object)

# S4 method for partition
p_attributes(.Object, p_attribute = NULL, ...)

# S4 method for partition
split(x, gap, ...)

is.partition(x)

# S4 method for partition
$(x, name)

# S4 method for partition,ANY
==(e1, e2)

# S4 method for partition,ANY
!=(e1, e2)

# S4 method for partition
%in%(x, table)

# S4 method for partition
zoom(x, ...)

# S4 method for partition
enrich(.Object, p_attribute = NULL, decode = TRUE,
  verbose = TRUE, mc = FALSE, ...)

# S4 method for partition
as.regions(x)

Arguments

object
A partition object.

.Object
A partition object.

p_attribute
a p-attribute (for enriching) / performing count.

...
further parameters passed into count when calling enrich, and ...

x
A partition object.

gap
An integer value specifying the minimum gap between regions for performing the split.

name
An s-attribute that will be assigned as key to a partition.

e1
A first expression, a partition object in this case.

e2
A second expression, the value of an s-attribute.

table
Values a s-attribute shall assume.

decode
logical value, whether to decode token ids into strings when performing count

verbose
logical value, whether to output messages

mc
logical or, if numeric, providing the number of cores

Details

As partition objects inherit from count and textstat class, methods available are view to inspect the table in the stat slot, name and name<- to retrieve/set the name of an object, and more.

The p_attributes-method returns the p-attributes defined for the corpus the partition is derived from, if argument p_attribute is NULL (the default). If p_attribute is defined, the unique values for the p-attribute are returned.

The split-method will split a partition object into a partition_bundle if gap between strucs exceeds a minimum number of tokens specified by 'gap'. Relevant to split up a plenary protocol into speeches. Note: To speed things up, the returned partitions will not include frequency lists. The lists can be prepared by applying enrich on the partition_bundle object that is returned.

The is.partition function returns a logical value whether x is a partition, or not.

The `$`-method will assign the argument name to the slot key and return the modified object.

The enrich-method will add a count of tokens defined by argument p_attribute to slot stat of the partition object.

Slots

name

A name to identify the object (character vector with length 1); useful when multiple partition objects are combined to a partition_bundle.

corpus

The CWB indexed corpus the partition is derived from (character vector with length 1).

encoding

Encoding of the corpus (character vector with length 1).

s_attributes

A named list with the s-attributes specifying the partition.

explanation

Object of class character, an explanation of the partition.

cpos

A matrix with left and right corpus positions defining regions (two columns).

annotations

Object of class list.

size

Total size of the partition (integer vector, length 1).

stat

An (optional) data.table with counts. If present, speeds up computation of cooccurrences, as count is already present.

metadata

Object of class data.frame, metadata information.

strucs

Object of class intger, the strucs defining the partition.

p_attribute

Object of class character indicating the p_attribute of the count in slot stat.

xml

Object of class character, whether the xml is flat or nested.

s_attribute_strucs

Object of class character the base node

key

Experimental, an s-attribute that is used as a key.

call

Object of class character the call that generated the partition

See also

Examples

m <- partition("GERMAPARLMINI", speaker = "Merkel", regex = TRUE)

#> ... get encoding: latin1

#> ... get cpos and strucs

m2 <- m$speaker
m2@key

#> [1] "speaker"

m <- partition("GERMAPARLMINI", speaker = "Merkel", regex = TRUE)

#> ... get encoding: latin1

#> ... get cpos and strucs

m$date == "2009-10-28"

#> ... Setting up partition 

#> ... getting cpos and strucs

#> ** partition object **
#> corpus:              GERMAPARLMINI 
#> name:                 
#> s-attributes:        speaker = Merkel 
#>                      date = 2009-10-28 
#> cpos:                5 pairs of corpus positions
#> size:                127 tokens
#> count:               not available

s <- partition("GERMAPARLMINI", interjection = "speech")

#> ... get encoding: latin1

#> ... get cpos and strucs

s$party != "NA"

#> ... Setting up partition 

#> ... getting cpos and strucs

#> ** partition object **
#> corpus:              GERMAPARLMINI 
#> name:                 
#> s-attributes:        interjection = speech 
#>                      party = CDU_CSU/SPD/FDP/DIE_LINKE/B90_DIE_GRUENEN 
#> cpos:                1818 pairs of corpus positions
#> size:                161164 tokens
#> count:               not available

s <- partition("GERMAPARLMINI", interjection = "speech")

#> ... get encoding: latin1

#> ... get cpos and strucs

s$date %in% c("2009-10-27", "2009-10-28")

#> ... Setting up partition 

#> ... getting cpos and strucs

#> ** partition object **
#> corpus:              GERMAPARLMINI 
#> name:                 
#> s-attributes:        interjection = speech 
#>                      date = 2009-10-27/2009-10-28 
#> cpos:                124 pairs of corpus positions
#> size:                10958 tokens
#> count:               not available

m <- partition("GERMAPARLMINI", speaker = "Merkel", regex = TRUE)

#> ... get encoding: latin1

#> ... get cpos and strucs

y <- zoom(m, date == "2009-10-28")

#> ... Setting up partition 

#> ... getting cpos and strucs

speeches <- partition("GERMAPARLMINI", interjection = "speech")

#> ... get encoding: latin1

#> ... get cpos and strucs

m <- zoom(speeches, date == "2009-10-28" & speaker == "Angela Dorothea Merkel")

#> ... Setting up partition 

#> ... getting cpos and strucs

#> ... Setting up partition 

#> ... getting cpos and strucs

not_unknown <- zoom(speeches, party != c("NA", "FDP"))

#> ... Setting up partition 

#> ... getting cpos and strucs

s_attributes(not_unknown, "party")

#> [1] "CDU_CSU"         "SPD"             "DIE_LINKE"       "B90_DIE_GRUENEN"