Partition class and methods.

S4 partition class and methods for instances of class partition.

# S4 method for partition
name(x)

# S4 method for partition,character
name(x) <- value

# S4 method for partition
length(x)

# S4 method for partition
hist(x, ...)

# S4 method for partition
length(x)

# S4 method for partition
aggregate(x)

# S4 method for partition
as.markdown(.Object, meta = getOption("polmineR.meta"),
  template = getTemplate(.Object), cpos = TRUE, cutoff = NULL,
  verbose = FALSE, ...)

# S4 method for partition
enrich(.Object, size = FALSE, pAttribute = NULL,
  id2str = TRUE, verbose = TRUE, mc = FALSE, ...)

# S4 method for partition
freq(.Object)

# S4 method for partition
highlight(.Object, html, highlight = list(),
  cqp = is.cqp, tooltips = NULL)

# S4 method for partition
pAttributes(.Object, pAttribute = NULL)

# S4 method for partition
view(.Object)

# S4 method for cooccurrences
view(.Object)

Arguments

x
a partition object

value
value

...
further parameters

.Object
a partition object

meta
...

template
template to use

cpos
...

cutoff
maximum number of tokens to decode

verbose
logical

size
logical

pAttribute
a p-attribute (for enriching)

id2str
logical

mc
logical or, if numeric, providing the number of cores

html
...

highlight
...

cqp
...

tooltips
...

Slots

name

Object of class "character" a name that may be useful

corpus

Object of class "character" the CWB corpus the partition is based on

encoding

Object of class "character" encoding of the corpus

sAttributes

Object of class "list" s-attributes specifying the partition

explanation

Object of class "character" an explanation of the partition

cpos

Object of class "matrix" corpus positions

annotations

Object of class "list"

pos

Object of class "list" with tables "abs", "rel" and "max"

size

Object of class "numeric" total size of the partition

metadata

Object of class "data.frame" metadata information

strucs

Object of class "numeric" the strucs defining the partition

pAttribute

Object of class "character" indicating the pAttribute of the count in slot tf

xml

Object of class "character" whether the xml is flat or nested

sAttributeStrucs

Object of class "character" the base node

call

Object of class "character" the call that generated the partition

See also

The partition-class inherits from the textstat-class, see respective documentation to learn more.