Reset Registry Directory.

A utility function to reset the environment variable CORPUS_REGISTRY. That may be necessary if you want use a CWB corpus that is not stored in the usual place. In particular, resetting the environment variable is required if you want to use a corpus delivered in a R package,

The polmineR package uses a subdirectory of the per-session temporary directory as a (temporary) registry. The registry function will return the path to this directory.

registry_reset(registryDir = registry(), verbose = TRUE)

registry()

Arguments

registryDir
path to the registry directory to be used

verbose
logical, whether to be verbose

Value

the registry directory used before resetting CORPUS_REGISTRY

Details

Resetting the CORPUS_REGISTRY environment variable is also necessary for the interface to CWB corpora.

To get the path to a package that contains a CWB corpus, use system.file (see examples).

See also

Examples

x <- system.file(package = "polmineR", "extdata", "cwb", "registry")
registry_reset(registryDir = x)

#> ... status: OK

registry()

#> [1] "/private/var/folders/r6/1k6mxnbj5077980k11xvr0q40000gn/T/RtmpFU8O3d/polmineR_registry"