Get s-attributes.

Structural annotations (s-attributes) of a corpus provide metainformation for regions of tokens. Gain access to the s-attributes available for a corpus or partition, or the values of s-attributes in a corpus/partition with the s_attributes-method.

s_attributes(.Object, ...)

# S4 method for character
s_attributes(.Object, s_attribute = NULL,
  unique = TRUE, regex = NULL, ...)

# S4 method for partition
s_attributes(.Object, s_attribute = NULL,
  unique = TRUE, ...)

Arguments

.Object
either a partition object or a character vector specifying a CWB corpus

...
to maintain backward compatibility, of argument sAttribute is used

s_attribute
name of a specific s-attribute

unique
logical, whether to return unique values only

regex
filter return value by applying a regex

Value

a character vector

Details

Importing XML into the Corpus Workbench (CWB) turns elements and element attributes into so-called s-attributes. There are two uses of the s_attributes-method: If the s_attribute parameter is NULL (default), the return value is a character vector with all s-attributes present in a corpus.

If s_attribute is the name of a specific s-attribute (a length 1 character vector), the values of the s-attributes available in the corpus/partition are returned.

If a character vector of s-attributes is provided, the method will return a data.table.

Examples

use("polmineR")

#> ... activating corpus: GERMAPARLMINI

#> ... activating corpus: REUTERS

s_attributes("GERMAPARLMINI")

#> [1] "interjection" "date"         "party"        "speaker"     

s_attributes("GERMAPARLMINI", "date") # dates of plenary meetings

#> [1] "2009-10-27" "2009-10-28" "2009-11-10" "2009-11-11" "2009-11-12"

P <- partition("GERMAPARLMINI", date = "2009-11-10")

#> ... get encoding: latin1

#> ... get cpos and strucs

s_attributes(P)

#> [1] "interjection" "date"         "party"        "speaker"     

s_attributes(P, "speaker") # get names of speakers

#>  [1] "Norbert Lammert"             "Angela Dorothea Merkel"     
#>  [3] "Frank-Walter Steinmeier"     "Birgit Homburger"           
#>  [5] "Wolfgang Thierse"            "Oskar Lafontaine"           
#>  [7] "Jürgen Trittin"              "Volker Kauder"              
#>  [9] "Joachim Poß"                 "Hans-Peter Friedrich"       
#> [11] "Agnes Krumwiede"             "Gerda Hasselfeldt"          
#> [13] "Arnold Vaatz"                "Guido Westerwelle"          
#> [15] "Gernot Erler"                "Andreas Schockenhoff"       
#> [17] "Jan  van Aken"               "Frithjof Schmidt"           
#> [19] "Dirk Niebel"                 "Sascha Raabe"               
#> [21] "Angelica Schwall-Düren"      "Karl-Theodor  zu Guttenberg"
#> [23] "Wolfgang Gehrcke"            "Omid Nouripour"             
#> [25] "Petra Pau"                   "Ilse Aigner"                
#> [27] "Waltraud Wolff"              "Hans-Michael Goldmann"      
#> [29] "Katrin Göring-Eckardt"       "Kirsten Tackmann"           
#> [31] "Ulrike Höfken"               "Peter Bleser"               
#> [33] "Elvira Drobinski-Weiß"       "Christel Happach-Kasan"     
#> [35] "Caren Lay"                   "Johannes Röring"            
#> [37] "Wilhelm Priesmeier"