Get s-attributes.

Structural annotations (s-attributes) of a corpus provide metainformation for regions of tokens. Gain access to the s-attributes available for a corpus or partition, or the values of s-attributes in a corpus/partition with the s_attributes-method.

s_attributes(.Object, ...)

# S4 method for character
s_attributes(.Object, s_attribute = NULL,
  unique = TRUE, regex = NULL, ...)

# S4 method for partition
s_attributes(.Object, s_attribute = NULL,
  unique = TRUE, ...)

Arguments

.Object
either a partition object or a character vector specifying a CWB corpus

...
to maintain backward compatibility, of argument sAttribute is used

s_attribute
name of a specific s-attribute

unique
logical, whether to return unique values only

regex
filter return value by applying a regex

Value

a character vector

Details

Importing XML into the Corpus Workbench (CWB) turns elements and element attributes into so-called s-attributes. There are two uses of the s_attributes-method: If the s_attribute parameter is NULL (default), the return value is a character vector with all s-attributes present in a corpus.

If s_attribute is the name of a specific s-attribute (a length 1 character vector), the values of the s-attributes available in the corpus/partition are returned.

If a character vector of s-attributes is provided, the method will return a data.table.

Examples

use("polmineR")

#> ... activating corpus: germaparlmini

#> ... activating corpus: reuters

s_attributes("GERMAPARLMINI")

#> [1] "interjection" "date"         "party"        "speaker"     

s_attributes("GERMAPARLMINI", "date") # dates of plenary meetings

#> Error in check_strucs(corpus = corpus, s_attribute = s_attribute, strucs = struc,     registry = registry): all values of vector strucs need to be >= 0

P <- partition("GERMAPARLMINI", date = "2009-11-10")

#> ... get encoding: latin1

#> ... get cpos and strucs

#> Error in rep(0L, times = maxAttr): ungültiges 'times' Argument

s_attributes(P)

#> Error in s_attributes(P): Objekt 'P' nicht gefunden

s_attributes(P, "speaker") # get names of speakers

#> Error in s_attributes(P, "speaker"): Objekt 'P' nicht gefunden