S4 textstat superclass.

The textstat-class (technically an S4 class) serves as a superclass for the classes features, context, and partition. Usually, the class will not be used directly. It offers a set of standard generic methods (such as head, tail, dim, nrow, colnames) its childs inherit. The core feature of textstat and its childs is a data.table in the slot stat for keeping data on text statistics of a corpus, or a partition.

# S4 method for textstat
name(x)

# S4 method for textstat,character
name(x) <- value

# S4 method for textstat
round(x, digits = 2L)

# S4 method for textstat
sort(x, by, decreasing = TRUE)

as.bundle(object, ...)

# S4 method for textstat,textstat
+(e1, e2)

# S4 method for textstat
subset(x, ...)

# S4 method for textstat
p_attributes(.Object)

# S4 method for textstat
view(.Object)

Arguments

x
A textstat object.

value
A character vector to assign as name to slot name of a textstat class object.

digits
no of digits

by
Column that will serve as the key for sorting.

decreasing
Logical, whether to return decreasing order.

object
a textstat object

...
Further arguments.

e1
A texstat object.

e2
Another texstat object.

.Object
A textstat object.

Details

A head-method will return the first rows of the data.table in the stat-slot. Use argument n to specify the number of rows.

A tail-method will return the last rows of the data.table in the stat-slot. Use argument n to specify the number of rows.

The methods dim, nrow and ncol will return information on the dimensions, the number of rows, or the number of columns of the data.table in the stat-slot, respectively.

Objects derived from the textstat class can be indexed with simple square brackets ("[") to get rows specified by an numeric/integer vector, and with double square brackets ("[[") to get specific columns from the data.table in the slot stat.

The colnames-method will return the column names of the data-table in the slot stat.

The methods as.data.table, and as.data.frame will extract the data.table in the slot stat as a data.table, or data.frame, respectively.

textstat objects can have a name, which can be retrieved, and set using the name-method and name<-, respectively.

Slots

p_attribute

Object of class character, p-attribute of the query.

corpus

A corpus specified by a length-one character vector.

stat

A data.table with statistical information.

name

The name of the object.

encoding

A length-one character vector, the encoding of the corpus.

Examples

use("polmineR")

#> ... activating corpus: germaparlmini

#> ... activating corpus: reuters

P <- partition("GERMAPARLMINI", date = ".*", p_attribute = "word", regex = TRUE)

#> ... get encoding: latin1

#> ... get cpos and strucs

#> Error in rep(0L, times = maxAttr): ungültiges 'times' Argument

y <- cooccurrences(P, query = "Arbeit")

#> Error in cooccurrences(P, query = "Arbeit"): Objekt 'P' nicht gefunden

# Standard generic methods known from data.frames work for objects inheriting
# from the textstat class

head(y)

#> Error in head(y): Objekt 'y' nicht gefunden

tail(y)

#> Error in tail(y): Objekt 'y' nicht gefunden

nrow(y)

#> Error in nrow(y): Objekt 'y' nicht gefunden

ncol(y)

#> Error in ncol(y): Objekt 'y' nicht gefunden

dim(y)

#> Error in eval(expr, envir, enclos): Objekt 'y' nicht gefunden

colnames(y)

#> Error in colnames(y): Objekt 'y' nicht gefunden

# Use brackets for indexing 

y[1:25]

#> Error in eval(expr, envir, enclos): Objekt 'y' nicht gefunden

y[,c("word", "ll")]

#> Error in eval(expr, envir, enclos): Objekt 'y' nicht gefunden

y[1:25, "word"]

#> Error in eval(expr, envir, enclos): Objekt 'y' nicht gefunden

y[1:25][["word"]]

#> Error in eval(expr, envir, enclos): Objekt 'y' nicht gefunden

y[which(y[["word"]] %in% c("Arbeit", "Sozial"))]

#> Error in eval(expr, envir, enclos): Objekt 'y' nicht gefunden

y[ y[["word"]] %in% c("Arbeit", "Sozial") ]

#> Error in eval(expr, envir, enclos): Objekt 'y' nicht gefunden